requirements.txt
whisper
langchain
pandas
matplotlib
python-dotenv
openai-whisper 
streamlit 
nltk